+++ addons/fleet/models/fleet_vehicle_trip_master.py
@@ -15,7 +15,7 @@ class FleetVehicleTripMaster(models.Model):
     name = fields.Char('Trip Name', required=True)
     distance = fields.Integer('Distance (km)', required=True)
     price = fields.Float('Unit Price(MYR)', required=True)
-    bucket = fields.Integer('Ton/bucket', required=True)
+    bucket = fields.Float('Ton/bucket', required=True)
     weight = fields.Integer('Default Weight (ton)', required=True)
     active = fields.Boolean(default=True)
